+++ linphone-app/ui/modules/Linphone/Contact/ContactMessageCounter.qml
@@ -11,10 +11,13 @@ Item {
   id: messageCounter
 
   property alias count: counterIcon.count
-  property bool isComposing
+  property var entry
+  property bool isComposing: Boolean(entry && entry.composers && entry.composers.length > 0)
+  property bool displayCounter: true
 
   implicitHeight: counterIcon.height + ContactMessageCounterStyle.verticalMargins * 2
   implicitWidth: counterIcon.width + ContactMessageCounterStyle.horizontalMargins * 2
+  visible: count > 0 ?(entry.unreadMessagesCount !== null || entry.missedCallsCount !== null) && displayCounter:false
 
   MessageCounter {
     id: counterIcon
@@ -27,6 +30,8 @@ Item {
       ? ('chat_is_composing_' + counterIcon.composingIndex)
       : 'chat_count'
     visible: messageCounter.count > 0 || messageCounter.isComposing
+    
+    count: messageCounter.entry?Number(messageCounter.entry.unreadMessagesCount) + Number(messageCounter.entry.missedCallsCount):0
 
     Timer {
       interval: 500