+++ linphone-app/ui/modules/Linphone/Styles/Contact/ContactMessageCounterStyle.qml
@@ -6,5 +6,5 @@ import QtQml 2.2
 
 QtObject {
   property int horizontalMargins: 0
-  property int verticalMargins: 10
+  property int verticalMargins: 5
 }